+++ src/useElapsedTime.ts
@@ -1,30 +1,60 @@
-import {
-  useState,
-  useRef,
-  useCallback,
-  useEffect,
-  useLayoutEffect,
-} from 'react'
-import type { Props, ReturnValue } from './types'
+import { useState, useRef, useCallback } from 'react'
+import { useIsomorphicEffect } from './useIsomorphicEffect'
 
-const useIsomorphicEffect =
-  typeof window === 'undefined' ? useEffect : useLayoutEffect
+type MayBe<T> = T | null
+
+export interface ReturnValue {
+  /** Current elapsed time in seconds */
+  elapsedTime: number
+  /** Reset method to reset the elapsed time and start over from the "startAt" value */
+  reset: () => void
+}
+
+export interface OnComplete {
+  /** Indicates if the loop should start over. Default: false */
+  shouldRepeat?: boolean
+  /** Delay in seconds before looping again. Default: 0 */
+  delay?: number
+}
+
+export interface Props {
+  /** Animation duration in seconds */
+  isPlaying: boolean
+  /** Animation duration in seconds */
+  duration?: number
+  /** Start the animation at provided time in seconds. Default: 0 */
+  startAt?: number
+  /** Update interval in seconds. Determines how often the component will rerender. When set to 0 the component will rerender on each key frame. Default: 0 */
+  updateInterval?: number
+  /** On animation complete event handler. It can be used to restart/repeat the animation by returning an object */
+  onComplete?: (totalElapsedTime: number) => OnComplete | void
+  /** On time update callback. It receives the current elapsedTime time in seconds*/
+  onUpdate?: (elapsedTime: number) => void
+}
 
 export const useElapsedTime = ({
   isPlaying,
   duration,
-  onComplete,
-  autoResetKey,
   startAt = 0,
+  updateInterval = 0,
+  onComplete,
+  onUpdate,
 }: Props): ReturnValue => {
   const [elapsedTime, setElapsedTime] = useState(startAt)
-  const totalElapsedTime = useRef(startAt * -1000) // keep in milliseconds to avoid summing up floating point numbers
-  const requestRef = useRef<number | null>(null)
-  const previousTimeRef = useRef<number | null>(null)
-  const repeatTimeoutRef = useRef<NodeJS.Timeout | null>(null)
-  const isMountedRef = useRef(false)
-  const durationRef = useRef(duration)
-  durationRef.current = duration
+  const totalElapsedTimeRef = useRef(startAt * -1000) // keep in milliseconds to avoid summing up floating point numbers
+  const requestRef = useRef<MayBe<number>>(null)
+  const previousTimeRef = useRef<MayBe<number>>(null)
+  const repeatTimeoutRef = useRef<MayBe<NodeJS.Timeout>>(null)
+  const loopRef = useRef({
+    elapsedTimeRef: elapsedTime,
+    durationRef: duration,
+    updateIntervalRef: updateInterval,
+  })
+  loopRef.current = {
+    ...loopRef.current,
+    durationRef: duration,
+    updateIntervalRef: updateInterval,
+  }
 
   const loop = (time: number) => {
     const timeSec = time / 1000
@@ -34,52 +64,55 @@ export const useElapsedTime = ({
       return
     }
 
-    let isCompleted = false
+    // get current elapsed time
+    const { durationRef, elapsedTimeRef, updateIntervalRef } = loopRef.current
     const deltaTime = timeSec - previousTimeRef.current
+    const currentElapsedTime = elapsedTimeRef + deltaTime
+
+    // update refs with the current elapsed time
     previousTimeRef.current = timeSec
+    loopRef.current = { ...loopRef.current, elapsedTimeRef: currentElapsedTime }
 
-    setElapsedTime((prevTime) => {
-      const currentElapsedTime = prevTime + deltaTime
-      isCompleted =
-        typeof durationRef.current === 'number' &&
-        currentElapsedTime >= durationRef.current
+    // set current display time
+    const currentDisplayTime =
+      updateIntervalRef === 0
+        ? currentElapsedTime
+        : (currentElapsedTime | 0) * updateIntervalRef
+    const isCompleted =
+      typeof durationRef === 'number' && currentDisplayTime >= durationRef
+    setElapsedTime(isCompleted ? durationRef! : currentDisplayTime)
 
-      return isCompleted ? durationRef.current! : currentElapsedTime
-    })
-
+    // repeat animation if not completed
     if (!isCompleted) {
       requestRef.current = requestAnimationFrame(loop)
     }
   }
 
-  // only for internal use
   const cleanup = () => {
     requestRef.current && cancelAnimationFrame(requestRef.current)
     repeatTimeoutRef.current && clearTimeout(repeatTimeoutRef.current)
     previousTimeRef.current = null
   }
 
-  const reset = useCallback(
-    (newStartAt?: number) => {
-      setElapsedTime(typeof newStartAt === 'number' ? newStartAt : startAt)
-      if (isPlaying) {
-        cleanup()
-        requestRef.current = requestAnimationFrame(loop)
-      }
-    },
-    [isPlaying]
-  )
-  useIsomorphicEffect(() => {
-    if (duration && elapsedTime >= duration) {
-      totalElapsedTime.current += duration * 1000
+  const reset = useCallback(() => {
+    setElapsedTime(startAt)
+    if (isPlaying) {
+      cleanup()
+      requestRef.current = requestAnimationFrame(loop)
+    }
+  }, [isPlaying])
 
-      const { shouldRepeat = false, delay = 0, newStartAt } =
-        onComplete?.(totalElapsedTime.current / 1000) || {}
+  useIsomorphicEffect(() => {
+    onUpdate?.(elapsedTime)
+
+    if (duration && elapsedTime >= duration) {
+      totalElapsedTimeRef.current += duration * 1000
+
+      const { shouldRepeat = false, delay = 0 } =
+        onComplete?.(totalElapsedTimeRef.current / 1000) || {}
 
       if (shouldRepeat) {
-        repeatTimeoutRef.current = setTimeout(() => {
-          reset(newStartAt)
-        }, delay * 1000)
+        repeatTimeoutRef.current = setTimeout(reset, delay * 1000)
       }
     }
   }, [elapsedTime, duration])
@@ -91,13 +124,6 @@ export const useElapsedTime = ({
 
     return cleanup
   }, [isPlaying])
-
-  // auto reset the animation when the autoResetKey changes
-  useIsomorphicEffect(() => {
-    isMountedRef.current && reset()
-    // set isMounted in the last effect
-    isMountedRef.current = true
-  }, [autoResetKey])
 
   return { elapsedTime, reset }
 }