+++ src/useIsomorphicEffect.ts
@@ -1,0 +1,4 @@
+import { useEffect, useLayoutEffect } from 'react'
+
+export const useIsomorphicEffect =
+  typeof window === 'undefined' ? useEffect : useLayoutEffect